Add helper to refine IntegrityErrors
When we normally catch an IntegrityError from Django, it can be hard to
determine exactly what the cause of the error was. Sometimes we assume
that the error was of one type, where the error actually had
another cause.

This utility will make it easy for us to tell the difference between
different types of IntegrityError, allowing us to be more specific about
which errors we handle.

Co-authored-by: Samuel Searles-Bryant <sam@samueljsb.co.uk>