Don't install unused requirements in CI
Before this change, we installed all the requirements used to run the
tests, when we only needed tox. Tox would then install those same
requirements in each of the virtual environments it created.

This change installs only tox in CI, trusting that it's the only package
required to run the tests.