Ensure PrimaryKey rules are for models with PKs
Before this change, it was possible to create a `PrimaryKey` rule for an
abstract model which doesn't have an explicit `.pk` field.

If we had tried to check for a match on that model, we would have seen
an `AssertionError`.

This change makes the `PrimaryKey` rule raise a `ModelHasNoPrimaryKey`
error when instantiated, so that the `assert` statement cannot trigger,
and is only of use to Mypy.