Check release type instead of branch when releasing
Before this change, we checked the branch name, but we're reacting to
pushed tags, so that did nothing.

This change switches over to checking that we're reacting to a tag. This
way we'll still be able to partially run this in PRs to check that the
build works, without making any attempt to push to PyPI.

Co-authored-by: Sam Searles-Bryant <sam.searles-bryant@kraken.tech>