Improve crash handling in type-ratchet script
Before this change, we were failing to catch crashes from mypy plugins,
which can cause mypy to crash with a return-code of 1.

Frustratingly, return-code 1 is what mypy returns when it finds type
errors.

This change tries to work out if mypy failed in an expected manner by
reading the last line of mypy's output and checking to see if it matches
the standard error format: "Found X errors ...".