Add Django and Psycopg2 to dev dependencies
We didn't install these before because they clashed with the
requirements of running Pytest in Tox. This caused issues where
developers and scripts expected these to be installed.

Now that we have isolated the dependencies of pytest-in-tox, we can
install these for local development, and in CI when running linters.

This causes the Mypy linter to begin failing. It reveals a
previously-hidden bug that I will address separately.