Change type signature of refine_integrity_error
We were using a mapping (ie: dict) from the matching rule to the refined
exception, it turned out that this was a pain to correctly type.

Seeing as we don't actually need a dict, we've changed this to a
sequence of 2-tuples, where the first element is the matching rule, and
the second is the exception.

Co-authored-by: Sam Searles-Bryant <sam.searles-bryant@kraken.tech>