Rework xfail test
Before this change, this test used Django's test transaction to catch
the failure outside the body of the test.

This was unreliable. The test passed CI even when the test body was
removed.

This change reworks the test so that the entirety of the tested logic is
inside the test.